Rename DEPLOY_DIR -> WHISK_DEPLOY_DIR
Renames the variable used to store the per-product deploy directory from
`DEPLOY_DIR_${product}` to `WHISK_DEPLOY_DIR_${product}`. The former is
a bad choice for the variable name because if any product has a name
that happens to correspond with a bitbake override, it will override the
`DEPLOY_DIR` variable, causing problems.

To remain compatible with existing deployments, the old
`DEPLOY_DIR_${product}` variable is still defined if `whisk.yaml` is
version `1`, but is removed starting with version `2`. The new
`WHISK_DEPLOY_DIR_${product}` variable is present in all config file
versions.