Add multiple Yocto version compatibility
In order to correctly handle the quirks of various different Yocto
versions (and importantly Yocto 4.0), a table of supported version with
their discrepancies has been added. A version in the whisk.yaml file can
specify which Yocto version it is compatible with in a number of ways:
 * Explicitly specifying the release name the version is compatible with
   in its "compat" property
 * Specifying "auto" for its "compat" property. Whisk will scan all
   includes layer.conf files until it finds LAYERSERIES_CORENAMES (set
   in OE-core) uses it as the compatible version
 * Omitting the "compat" property, in which case it defaults to "auto"

In addition, a new property called `env_passthrough_vars` is added to
the `hooks` section. This controls what environment variables from the
host are passed in to bitbake; this is required so that the correct
environment variable can be used to pass these variables along based on
the compat version.

Finally, whisk now takes responsibility for setting CONF_VERSION in the
config file it writes out, so if you were setting that manually before
it needs to be removed.